feat(user-auth): implement logout

+++ lib/presentation/auth/bloc/auth_bloc.dart
@@ -1,10 +1,6 @@
-import 'dart:convert';
-
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
-import 'package:tomacare/domain/entities/auth.dart';
-// import 'package:tomacare/domain/entities/auth.dart';
 import 'package:tomacare/service/auth_service.dart';
 import 'package:tomacare/service/save_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
@@ -21,25 +17,32 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
         final String? token = await saveService.getToken();
 
         if (token == null) {
-          print('null');
           emit(AuthFailed('perlu login'));
         } else {
           emit(AuthSucess(token));
         }
       } catch (e) {
-        print(e.toString());
         emit(AuthFailed(e.toString()));
       }
     });
 
     on<Logout>((event, emit) async {
-      saveService.deleteToken();
-      emit(AuthFailed('logout'));
+      emit(AuthLoading(isLoading: true));
+      try {
+        final String? token = await saveService.getToken();
+
+        final res = await authService.logout(token!);
+        if (res['success']) {
+          saveService.deleteToken();
+          emit(AuthFailed('logout'));
+        }
+        emit(AuthFailed("Gagal Logout!"));
+      } catch (e) {
+        emit(AuthFailed(e.toString()));
+      }
     });
 
     on<Register>((event, emit) async {
-      print('start register');
-
       try {
         final Map<String, dynamic> response = await authService.register(
             email: event.email,
@@ -65,29 +68,30 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
       }
     });
 
-    on<SignInWithGoogle>((event, emit) async {
-      emit(AuthLoading(isLoading: true));
-      final GoogleSignIn googleSignIn = GoogleSignIn();
-      String? response;
-      try {
-        final account = await googleSignIn.signIn();
-        if (account != null) {
-          final GoogleSignInAuthentication auth = await account.authentication;
-          final accessToken = auth.accessToken;
-          if (accessToken != null){
-            response = await authService.googleLogin(googleAccessToken: accessToken);
-          }
+    on<SignInWithGoogle>(
+      (event, emit) async {
+        emit(AuthLoading(isLoading: true));
+        final GoogleSignIn googleSignIn = GoogleSignIn();
 
-          if (response != null) {
-            saveService.storeToken(response);
-            emit(AuthSucess(response));
-          } else {
-            emit(AuthFailed("gagal login"));
+        try {
+          final account = await googleSignIn.signIn();
+          if (account != null) {
+            final GoogleSignInAuthentication auth =
+                await account.authentication;
+            final accessToken = auth.accessToken;
+            if (accessToken != null) {
+              final response =
+                  await authService.googleLogin(googleAccessToken: accessToken);
+              saveService.storeToken(response);
+              emit(AuthSucess(response['access_token']));
+            } else {
+              emit(AuthFailed("gagal login"));
+            }
           }
+        } catch (e) {
+          emit(AuthFailed("gagal login"));
         }
-      } catch (e) {
-        emit(AuthFailed("gagal login"));
-      }
-    },);
+      },
+    );
   }
 }
